Allow setting ControlAddr
ControlAddr is useful in case it's a different hostname than the short
name defined in ControlMachine. We need it because our service nodes
have have the same short names on both the external and internal
interface, and by default the login node tries to connect to the
control machine over the external interface which doesn't work due to
the firewall.