Send a slurm efficiency report if the user requests mail notification
The seff script is a simple script which checks the CPU and memory
usage vs. allocated. smail is a small wrapper around /bin/mail which
adds the seff output to the body of the notification emails that slurm
sends (by default the body is empty). Both scripts are part of the
slurm-seff package.

See https://bugs.schedmd.com/show_bug.cgi?id=1611