support for SLURM 17.02 - #77
With SLURM 17.02 we need to do some tasks differently:

 - install slurm-contribs instead of slurm-sjstat,sjobexit,seff
 - no longer install slurm-slurmdb-direct
 - no longer disable slurm initd script (as it's not there anymore)
 - change fgci_slurmrepo_version default from pointing to yum repo
with 15.08 to one with 17.02

We do these by default, but if fgci_slurmrepo_version is set to a number
lower than 1702 then we do things the old way.

With this commit we now also test SLURM 15.08, 16.05 and 17.02 on every
commit.

Added a bit of documentation and comments about testing